+++ src/modules/url/url.service.ts
@@ -1,10 +1,12 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import { BadRequestException, HttpStatus, Injectable } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { CreateUrlDto } from './dto/create-url.dto';
 import { generateShortCode } from '../../helpers/code-generator';
 import { UrlRepository } from './url.repository';
 import { UpdateUrlDto } from './dto/update-url.dto';
 import { GetUrlContract } from './contract/url.contract';
+import { BaseResponseDTO } from '../../common/dto/response.dto';
+import { FormattedUrl } from './interface/url.interface';
 
 @Injectable()
 export class UrlService {
@@ -17,7 +19,7 @@ export class UrlService {
     this.BASE_URL = configService.getOrThrow<string>('BASE_URL');
   }
 
-  async incrementAccessCount({ short_code }: GetUrlContract): Promise<void> {
+  async incrementAccessCount({ short_code }: GetUrlContract) {
     if (!short_code) {
       throw new Error('Short code is required to increment access count.');
     }
@@ -25,22 +27,35 @@ export class UrlService {
     await this.urlRepository.incrementAccessCount({ short_code });
   }
 
-  //#region  GETS
-  async getOriginalUrl(short_code: string): Promise<string | null> {
+  //#region GETS
+  async getOriginalUrl(short_code: string): Promise<BaseResponseDTO<string>> {
     if (!short_code) {
       throw new BadRequestException(
         'Short code is required to retrieve the original URL.',
       );
     }
 
-    return await this.urlRepository.getOriginalUrl({ short_code });
+    const originalUrl = await this.urlRepository.getOriginalUrl({ short_code });
+
+    if (!originalUrl) {
+      throw new BadRequestException('Short URL not found.');
+    }
+
+    return new BaseResponseDTO({
+      data: originalUrl,
+      message: 'Original URL retrieved',
+    });
   }
 
-  async getUrlsByUser({ user_id }: { user_id: string }) {
+  async getUrlsByUser({
+    user_id,
+  }: {
+    user_id: string;
+  }): Promise<BaseResponseDTO<FormattedUrl[]>> {
     const urls = await this.urlRepository.getUrlsByUser(user_id);
 
     const formattedUrls = urls.map((url) => ({
-      irl_id: url.url_id,
+      url_id: url.url_id,
       original_url: url.original,
       short_url: `${this.BASE_URL}/${url.short_code}`,
       click_count: url.click_count,
@@ -48,14 +63,15 @@ export class UrlService {
       updated_at: url.updated_at,
     }));
 
-    return formattedUrls;
+    return new BaseResponseDTO({
+      data: formattedUrls,
+      message: 'User URLs retrieved successfully',
+    });
   }
-
   //#endregion GETS
 
   //#region POSTS
-
-  async shortenUrl(payload: CreateUrlDto) {
+  async shortenUrl(payload: CreateUrlDto): Promise<BaseResponseDTO<string>> {
     const shortCode = generateShortCode(6);
 
     const url = await this.urlRepository.shortenUrl({
@@ -64,25 +80,33 @@ export class UrlService {
       user_id: payload.user_id,
     });
 
-    return {
-      shortUrl: `${this.BASE_URL}/${url.short_code}`,
-    };
+    return new BaseResponseDTO({
+      data: `${this.BASE_URL}/${url.short_code}`,
+      message: 'URL shortened successfully',
+    });
   }
-
   //#endregion POSTS
 
   //#region PUTS
-  async updateUrl(payload: UpdateUrlDto) {
+  async updateUrl(
+    payload: UpdateUrlDto,
+  ): Promise<BaseResponseDTO<FormattedUrl>> {
     const urlExists = await this.urlRepository.findUrlById(payload.url_id);
 
     if (!urlExists) {
       throw new Error(`URL with ID ${payload.url_id} does not exist.`);
     }
 
+    if (urlExists?.user_id !== payload.user_id) {
+      throw new BadRequestException(
+        'You do not have permission to delete this URL.',
+      );
+    }
+
     const url = await this.urlRepository.updateUrl(payload);
 
     const formatted = {
-      irl_id: url.url_id,
+      url_id: url.url_id,
       original_url: url.original,
       short_url: `${this.BASE_URL}/${url.short_code}`,
       click_count: url.click_count,
@@ -90,20 +114,38 @@ export class UrlService {
       updated_at: url.updated_at,
     };
 
-    return formatted;
+    return new BaseResponseDTO({
+      data: formatted,
+      message: 'URL updated successfully',
+    });
   }
   //#endregion PUTS
 
   //#region DELETES
-  async deleteUrl({ url_id }: { url_id: string }) {
+  async deleteUrl({
+    url_id,
+    user_id,
+  }: {
+    url_id: string;
+    user_id?: string;
+  }): Promise<BaseResponseDTO> {
     const urlExists = await this.urlRepository.findUrlById(url_id);
 
     if (!urlExists) {
       throw new Error(`URL with ID ${url_id} does not exist.`);
     }
 
-    return this.urlRepository.deleteUrl(url_id);
-  }
+    if (urlExists?.user_id !== user_id) {
+      throw new BadRequestException(
+        'You do not have permission to delete this URL.',
+      );
+    }
 
+    await this.urlRepository.deleteUrl(url_id);
+    return new BaseResponseDTO({
+      message: 'URL deleted successfully',
+      code: HttpStatus.NO_CONTENT,
+    });
+  }
   //#endregion DELETES
 }
